fixed generation video glitch, add fuel low sound

+++ src/playarea.h
@@ -13,8 +13,8 @@ struct PlayBlock
 
 struct PlayArea
 {
-    int16_t m_x, m_y;
-    int16_t m_previousx, m_previousy;
+    int32_t m_x, m_y;
+    int32_t m_previousx, m_previousy;
     int16_t m_changedy;
     int16_t m_offsetX, m_offsetY;
     uint16_t m_width, m_height;
